fix(api): use PostgreSQL repository for proxy toggle and quota actions
Account action handlers (toggle_proxy, warmup, refresh_quota) called
file-based load_account() which fails with 404 when accounts come from
PostgreSQL (UUIDs don't match JSON filenames). Now uses repo.get_account()
when PG is available, with file-only fallback.

+++ crates/antigravity-core/src/proxy/mappers/openai/request/generation_config.rs
@@ -69,7 +69,7 @@ pub fn build_generation_config(
 
             // maxOutputTokens calculation (same as original logic but with resolved budget)
             if let Some(max_tokens) = request.max_tokens {
-                if (max_tokens as i64) <= budget as i64 {
+                if u64::from(max_tokens) <= budget {
                     gen_config["maxOutputTokens"] = json!(budget + THINKING_MIN_OVERHEAD);
                 } else {
                     gen_config["maxOutputTokens"] = json!(max_tokens);
